Move zonelist to its own file
